feat: refactor api response descriptions and schema definitions across multiple OAR test cases
- Updated response descriptions to clarify missing headers in OAR053 and OAR114.
- Changed schema definitions for parameters in OAR025 to include "schema" property.
- Adjusted format for numeric properties in OAR016 to use "int32" instead of "decimal".
- Enhanced response schemas in OAR034 to include content type specifications.
- Removed redundant components in OAR053 and streamlined response structures.
- Ensured consistent header definitions across various endpoints in OAR114.

+++ test/oap3/format/OAR016/fail-example.js
@@ -17,7 +17,7 @@ module.exports = {
                   "properties": {
                     "amount": {
                       "type": "number",
-                      "format": "decimal"
+                      "format": "int32"
                     },
                     "nested": {
                       "type": "object",
